refactor: split cli options into seperate modules
